Convert geo mappings and useUserLocation utils to TypeScript
